Add repeated code that checks session and user_id.
This can be tested by creating users in the database and "logging in"
using the username. Since the example is simplified, there are no
passwords.Then, create articles in the database that correspond to
those users.  After logging in, you should only see articles that belong
to the user that is logged in.